output: Extract file review change segmentation
File review model building already separates layout, actionable selection
derivation, pagination, and ReviewChange record assembly.

The change builder still owns row segmentation and record construction in
one function, making the displayability split and actionable grouping logic
harder to inspect independently.

This commit moves row segmentation into file_review_change_segments and
keeps file_review_changes focused on constructing ReviewChange records from
prepared segments.

Subsequent commits will add direct segment coverage and an import-boundary
guard.
